Week4/ Web API
I faced a lot of difficulties while performing this code.
Firstly,I had an issue with .NET 7 version which lead to several feature errors, so I upgraded it to .NET 9 version.

Secondly, I saw that I had to set in the object initializer in the Department.Name

Thirdly, I was not able to use Swagger properly since it was not loading due to some misconfigured package references I did in the 1st issue.

Fourthly, I spent my maximum time in rectifying the JWT and CORS authentication issue. I was not able to generate JWT tokens correctly, so configured CORS before authentication and authorization.

Although, it was nerve-racking, but I still enjoyed doing this project. I would like to do something like this all over again.